+++ Controllers/KeeperController.php
@@ -3,9 +3,9 @@
 namespace Controllers;
 
 
-use DAO\KeeperSQLDAO as KeeperDAO;
-use DAO\ReservationSQLDAO as ReservationDAO;
-use DAO\ReviewsSQLDAO as ReviewsDAO;
+use DAO\SQLDAO\KeeperDAO as KeeperDAO;
+use DAO\SQLDAO\ReservationDAO as ReservationDAO;
+use DAO\SQLDAO\ReviewsDAO as ReviewsDAO;
 use Exception;
 use Models\Keeper as Keeper;
 use Models\ReservationState as ReservationState;
